Reference

NODEData.LargeNODEDataloader

NODEData.NODEDataloader

NODEData.get_trajectory

NODEData.gpuoff

NODEData.gpuon

NODEData.LargeNODEDataloader — Type

LargeNODEDataloader(sol, N_batch, N_length, name, base_path=""; dt=nothing, valid_set=nothing)

Impelements an array of NODEDataloader that are too large for the RAM and are thus saved to the hard drive (temporally). They can be loaded by just indexing the

NODEData.NODEDataloader — Type

NODEDataloader{T,N} <: AbstractNODEDataloader{T,N}

Struct containing batched data for sequence learning of ODEs. Can be indexed and interated over. Each batch returns a tuple (t, data(t)).

Initialized with

NODEData(sol::SciMLBase.AbstractTimeseriesSolution, N_length::Integer; dt=nothing, valid_set=nothing)

sol: DE solution

N_length: length of each batch

dt: time increment that sol is interpolated at. If nothing then the sol.t is used as the time steps of the data

valid_set if validset ∈ [0,1] splits the data into a train and valid set with `validset` of the share of the data belonging to the valid_set.

NODEDataloader(data::AbstractArray{T,N}, t::AbstractArray{T,1}, N_length::Integer)

data:: Data that is already in Ndim1 x ... x N_t format

t: time axis

N_length: length of each batch

NODEData.get_trajectory — Method

get_trajectory(data::NODEDataloader, N)

Returns a (t, x(t)) tuple of length N.

NODEData.gpuoff — Method

gpuoff()

Manually toggle GPU use off

NODEData.gpuon — Method

gpuon()

Manually toggle GPU use on (if available)