Reference

NODEData.LargeNODEDataloader

NODEData.NODEDataloader

NODEData.SingleTrajectoryBatchedOSADataloader

NODEData.MultiTrajectoryBatchedNODEDataloader

NODEData.detect_sol_array_type

NODEData.get_trajectory

NODEData.get_trajectory

NODEData.gpuoff

NODEData.gpuon

NODEData.single_trajectory_from_batched

NODEData.LargeNODEDataloader — Type

LargeNODEDataloader(sol, N_batch, N_length, name, base_path=""; dt=nothing, valid_set=nothing)

Impelements an array of NODEDataloader that are too large for the RAM and are thus saved to the hard drive (temporally). They can be loaded by just indexing the

NODEData.NODEDataloader — Type

NODEDataloader{T,N} <: AbstractNODEDataloader{T,N}

Struct containing batched data for sequence learning of ODEs. Can be indexed and interated over. Each batch returns a tuple (t, data(t)).

Initialized with

NODEData(sol::SciMLBase.AbstractTimeseriesSolution, N_length::Integer; dt=nothing, valid_set=nothing, GPU=nothing)

sol: DE solution

N_length: length of each batch

dt: time increment that sol is interpolated at. If nothing then the sol.t is used as the time steps of the data

valid_set if validset ∈ [0,1] splits the data into a train and valid set with `validset` of the share of the data belonging to the valid_set.

GPU if nothing the output is automatically chosen to be on GPU/CPU based on sol, if GPU==true or GPU==false the automatic choice is overwritten

NODEDataloader(data::AbstractArray{T,N}, t::AbstractArray{T,1}, N_length::Integer)

data:: Data that is already in Ndim1 x ... x N_t format

t: time axis

N_length: length of each batch

NODEData.SingleTrajectoryBatchedOSADataloader — Method

SingleTrajectoryBatchedOSADataloader(data::AbstractArray{T,N}, t::AbstractArray{U,1}, N_batch::Int=1; valid_set=nothing, GPU::Union{Bool, Nothing}=nothing) where {T,U,N}

Prepares a single trajectory batches one-step-ahead dataloader. When indexed it returns tuples (t, x) which are batches of 2-element trajectories, so that they contain the initial condition and one step ahead of the dynamical system. The batches are in format (Ndims ... x Nbatch x N_t). t contains the time information of each of these trajectory snippets.

Inputs

data: Trajectory (Ndims .... x Nt)

t: time steps of the trajectory

NODEData.MultiTrajectoryBatchedNODEDataloader — Method

MultiTrajectoryBatchedNODEDataloader(trajectories::AbstractVector, N_length::Integer; valid_set=nothing, GPU::Union{Bool, Nothing}=nothing) where {T,U,N}

trajectories is a Vector of trajectories/tuples (t, x(t)), for the batched NODEDataloader it is assumed that all t are equal, and only the first one is used. The constructor enforces the equal length, but only warns about non-equal elements

Actually the routine just wraps around NODEDataloader but concatanates the trajectories in such a way that it works with it.

NODEData.detect_sol_array_type — Method

detect_sol_array_type(sol::Union{SciMLBase.AbstractTimeseriesSolution, SciMLBase.AbstractDiffEqArray})

Returns true if the solution is on a GPU, false if it is not CPU.

NODEData.get_trajectory — Method

get_trajectory(data::NODEDataloader, N)

Returns a (t, x(t)) tuple of length N.

NODEData.get_trajectory — Method

get_trajectory(data::SingleTrajectoryBatchedOSADataloader{T,U,N}, N; N_batch=0)

Returns a (t, x(t)) tuple of length N. If N_batch != 0 the trajectory is repeated N_batch along an additional batch axis.

NODEData.gpuoff — Method

gpuoff()

Manually toggle GPU use off

NODEData.gpuon — Method

gpuon()

Manually toggle GPU use on (if available)

NODEData.single_trajectory_from_batched — Method

single_trajectory_from_batched(data::NODEDataloader, N::Int)

Returns a single trajectory (t,x(t)) fomr the dataloader but repeats it along the N_batch dimension. Useful for testing the performance of methods that can't change the batch size dynamically.